Move compilation into Module from Instance.
