+++ crates/api/src/instance.rs
@@ -6,9 +6,7 @@ use crate::trampoline::take_api_trap;
 use crate::trap::Trap;
 use crate::types::{ExportType, ExternType};
 use anyhow::{Error, Result};
-use std::cell::RefCell;
-use std::collections::{HashMap, HashSet};
-use std::rc::Rc;
+use std::collections::HashSet;
 use wasmtime_jit::{CompiledModule, Resolver};
 use wasmtime_runtime::{Export, InstanceHandle, InstantiationError};
 
@@ -25,39 +23,23 @@ impl Resolver for SimpleResolver<'_> {
 }
 
 fn instantiate_in_context(
-    store: &Store,
-    data: &[u8],
+    compiled_module: &CompiledModule,
     imports: &[Extern],
-    module_name: Option<&str>,
     context: Context,
-    exports: Rc<RefCell<HashMap<String, Option<wasmtime_runtime::Export>>>>,
 ) -> Result<(InstanceHandle, HashSet<Context>), Error> {
-    let mut contexts = HashSet::new();
-    let debug_info = context.debug_info();
     let mut resolver = SimpleResolver { imports };
-    let mut compiled_module = CompiledModule::new(
-        &mut context.compiler(),
-        data,
-        module_name,
-        &mut resolver,
-        exports,
-        debug_info,
-    )?;
-
-    // Register all module signatures
-    for signature in compiled_module.module().signatures.values() {
-        store.register_wasmtime_signature(signature);
-    }
-
-    let instance = compiled_module.instantiate().map_err(|e| -> Error {
-        if let Some(trap) = take_api_trap() {
-            trap.into()
-        } else if let InstantiationError::StartTrap(msg) = e {
-            Trap::new(msg).into()
-        } else {
-            e.into()
-        }
-    })?;
+    let instance = compiled_module
+        .instantiate(&mut resolver)
+        .map_err(|e| -> Error {
+            if let Some(trap) = take_api_trap() {
+                trap.into()
+            } else if let InstantiationError::StartTrap(msg) = e {
+                Trap::new(msg).into()
+            } else {
+                e.into()
+            }
+        })?;
+    let mut contexts = HashSet::new();
     contexts.insert(context);
     Ok((instance, contexts))
 }
@@ -78,15 +60,8 @@ impl Instance {
     pub fn new(module: &Module, externs: &[Extern]) -> Result<Instance, Error> {
         let store = module.store();
         let context = store.context().clone();
-        let exports = store.global_exports().clone();
-        let (mut instance_handle, contexts) = instantiate_in_context(
-            module.store(),
-            module.binary().expect("binary"),
-            externs,
-            module.name(),
-            context,
-            exports,
-        )?;
+        let (mut instance_handle, contexts) =
+            instantiate_in_context(&*module.compiled_module(store)?, externs, context)?;
 
         let exports = {
             let mut exports = Vec::with_capacity(module.exports().len());
