+++ crates/jit/src/lib.rs
@@ -26,6 +26,7 @@ mod code_memory;
 mod compiler;
 mod context;
 mod function_table;
+mod imports;
 mod instantiate;
 mod link;
 mod namespace;
